+++ CHANGELOG.md
@@ -34,7 +34,9 @@
 - **(!)** `DOMImporter` (`HTMLImporter`/`XMLImporter`) takes two mandatory configuration parameters: `schema` and `converters`
 
   If you have used the `Configurator` API to create importers and exporters, you should be good.
-- **(!)** removed some API we don't want to support anymore:
+- **(!)** DEPRECATIONS:
+  - `DocumentChange.isAffected()`: Use `DocumentChange.hasUpdated()` instead
+- **(!)** removed some API which we don't want to support anymore:
   - `DOMElement.getRoot()`: use `el.getOwnerDocument()` instead
   - `Selection.getFragments()`: implementation was overly complicated.
     If you want to iterate nodes of a container selection use `sel.getNodeIds()`